requirements.txt
# Stuff easy to break with updates
torch
torchvision
pytorch-lightning
hydra-core
wandb
streamlit
# hydra-joblib-launcher

# Stable stuff usually backward compatible
dvc
python-dotenv
matplotlib
stqdm
